Create RFID-MFRC522.py
Installing MicroPython MFRC522 Library
For this project we will require the MicroPython MFRC522 library. Copy this library and save it in your Raspberry Pi Pico with the respective file name (RFID-MFRC522.py) from the GitHub link.

Open a new file in Thonny. Copy the library given below or from the link given above. Save it to Raspberry Pi Pico with the name RFID-MFRC522.py under the lib folder.